adding classes
Adding outlines for the classes we will be using our algorithm implementations on. These classes aren't finished yet, they're only being shared so that collaborators can see how I plan on implementing the classes. This will also give them the ability to make suggestions regarding what I should or should not do when writing the implementation. I'll finish the implementations tomorrow and work on some of the algorithms later (and maybe even write some client code for the program)