Adding main.py
Here's the client code I've written to be used for the demonstration. It could use more work and fix. The code also can't use gaussian elimination and RREF as of right now. I'll work on the code later since I'm just extremely tired right now and want to sleep. Some help on